Increase ControlPersist time
"Without ControlPersist, once the first SSH session you open is closed,
all other sessions on that connection are closed as well.
This can lead to a variety of surprising behavior: if you exit the
initial SSH shell while other sessions are sharing the connection,
the process will just hang. If you send it a Ctrl-C, all of those other
sessions will be abruptly terminated."

From:
https://developer.rackspace.com/blog/speeding-up-ssh-session-creation/